requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
python-telegram-bot>=20.7
schedule>=1.2.0
python-dotenv>=1.0.0
requests>=2.31.0
